+++ packages/stream_video_flutter/lib/src/theme/components/participant_label_theme.g.theme.dart
@@ -156,6 +156,7 @@ mixin _$StreamParticipantLabelStyle {
         t,
       ),
       showAudioIndicator: t < 0.5 ? a.showAudioIndicator : b.showAudioIndicator,
+      showVideoOffIcon: t < 0.5 ? a.showVideoOffIcon : b.showVideoOffIcon,
     );
   }
 
@@ -181,6 +182,7 @@ mixin _$StreamParticipantLabelStyle {
     double? microphoneIconSize,
     Color? microphoneOffColor,
     bool? showAudioIndicator,
+    bool? showVideoOffIcon,
   }) {
     final _this = (this as StreamParticipantLabelStyle);
 
@@ -209,6 +211,7 @@ mixin _$StreamParticipantLabelStyle {
       microphoneIconSize: microphoneIconSize ?? _this.microphoneIconSize,
       microphoneOffColor: microphoneOffColor ?? _this.microphoneOffColor,
       showAudioIndicator: showAudioIndicator ?? _this.showAudioIndicator,
+      showVideoOffIcon: showVideoOffIcon ?? _this.showVideoOffIcon,
     );
   }
 
@@ -247,6 +250,7 @@ mixin _$StreamParticipantLabelStyle {
       microphoneIconSize: other.microphoneIconSize,
       microphoneOffColor: other.microphoneOffColor,
       showAudioIndicator: other.showAudioIndicator,
+      showVideoOffIcon: other.showVideoOffIcon,
     );
   }
 
@@ -284,7 +288,8 @@ mixin _$StreamParticipantLabelStyle {
         _other.speakingColor == _this.speakingColor &&
         _other.microphoneIconSize == _this.microphoneIconSize &&
         _other.microphoneOffColor == _this.microphoneOffColor &&
-        _other.showAudioIndicator == _this.showAudioIndicator;
+        _other.showAudioIndicator == _this.showAudioIndicator &&
+        _other.showVideoOffIcon == _this.showVideoOffIcon;
   }
 
   @override
@@ -314,6 +319,7 @@ mixin _$StreamParticipantLabelStyle {
       _this.microphoneIconSize,
       _this.microphoneOffColor,
       _this.showAudioIndicator,
+      _this.showVideoOffIcon,
     ]);
   }
 }